scrape.js: Detect if GCI name is GitHub username
This detects when a student's GCI display name is their GitHub
username. To ensure we don't get the wrong account, this will first
verify that the user has been active since the start of GCI. If they
have, we then check if the user has created an issue or PR to the
organization's GitHub org. If so, we consider them the correct user.

Closes https://github.com/coala/gci-leaders/issues/32