scrape.js: Reduce number of API calls
Once we get search results from GitHub, be it an organization account
or a student account, we know it will almost certainly stay constant.
This will use the data from previous builds as a base to build off of
for future builds. Once we get data for a certain organization or
user, that data won't be searched for again.

Closes https://github.com/coala/gci-leaders/issues/27