scrape.js: Handle username changes and deletions
Prior to this, if a user's GitHub account was cached, it would be
assumed to be correct for all future builds. However, this is not
the case if the user has changed or deleted their account. This
will verify that the cached username exists before using it, and
rerun the username finding code if it does not exist.

Closes https://github.com/coala/gci-leaders/issues/93